pyproject.toml
[project]
name = "breeze"
version = "0.1.0"
description = "High-performance MCP server for semantic code search and indexing"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "lancedb>=0.5.0",
    "fastmcp>=2",
    "sentence-transformers>=4",
    "pydantic>=2.0.0",
    "numpy",
    "polars",
    "pyarrow",
    "orjson",
    "uvicorn",
    "anyio",
    "einops",
    "platformdirs>=4",
    "typer>=0.16.0",
    "rich>=14.0.0",
    "watchdog>=6.0.0",
    "aiofiles>=24.1.0",
    "python-magic>=0.4.27",
    "identify>=2.6.12",
    "gitignore_parser>=0.1.12",
    "tree-sitter-language-pack>=0.1.3",
    "voyageai>=0.3.2",
    "openai>=1.84.0",
    "cohere>=5.15.0",
    "google-generativeai>=0.8.5",
    "python-dotenv",
    "tokenizers>=0.21.0",
    "tenacity>=9",
    "uuid-utils>=0.11.0",
]

[project.scripts]
breeze = "breeze.cli:main"

[dependency-groups]
dev = ["ruff>=0.11.12", "pytest>=8.0.0", "pytest-asyncio>=0.23.0"]
